putting filter out of the uniform type structure
To have a convertibility between  locally and [filter of _],
and to be able to simplify filter_of on the fly.

This introduces a potential discrepency between product of filters and the filter on a product :-/

There is also a running error : cannot infer a filter typeclass on locally of a prod... why???
Recurring fix:
Unshelve.
eapply @filter_filter.
exact: locally_filter.
Qed.